+++ cksum.h
@@ -3,5 +3,5 @@
  * All Rights Reserved.  See COPYRIGHT.
  */
 
-ssize_t do_cksum( char *path, char *cksum_b64 );
-ssize_t do_acksum( char *path, char *cksum_b64, struct applefileinfo *afinfo );
+off_t do_cksum( char *path, char *cksum_b64 );
+off_t do_acksum( char *path, char *cksum_b64, struct applefileinfo *afinfo );
