+++ connect.h
@@ -7,12 +7,12 @@ SNET * connectsn( char *host, int port );
 int closesn( SNET *sn );
 
 int retr( SNET *sn, char *pathdesc, char *path, char *temppath,
-    ssize_t transize, char *trancksum );
+    off_t transize, char *trancksum );
 int retr_applefile( SNET *sn, char *pathdesc, char *path, char *temppath,
-    ssize_t transize, char *trancksum );
+    off_t transize, char *trancksum );
 
 int n_stor_file( SNET *sn, char *pathdesc, char *path );
-int stor_file( SNET *sn, char *pathdesc, char *path, size_t transize,
+int stor_file( SNET *sn, char *pathdesc, char *path, off_t transize,
     char *trancksum );
-int stor_applefile( SNET *sn, char *pathdesc, char *path, size_t transize,
+int stor_applefile( SNET *sn, char *pathdesc, char *path, off_t transize,
     char *trancksum, struct applefileinfo *afinfo );
